requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==3.2.15
django-modeltranslation==0.18.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.15.1
Faker==14.1.0
inflection==0.5.1
jsonschema==4.14.0
mccabe==0.6.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
celery
django_celery_beat==2.4.0